src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "beanie>=2.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
